Adapt the tests
