+++ src/tests/mod.rs
@@ -1,0 +1,7 @@
+#[allow(dead_code)]
+pub mod utils;
+
+mod read;
+mod write;
+mod network;
+mod parametric;
