+++ src/tests/utils/mod.rs
@@ -1,4 +1,6 @@
 #![allow(dead_code)]
+use std::{path::PathBuf, fs::File};
+
 use hex_literal::hex;
 use sha2::Digest;
 
@@ -36,4 +38,12 @@ pub fn assert_checksum_samedata<const N: usize>(buf: &[u8], data: u8) {
     let mut hasher = sha2::Sha256::new();
     hasher.update(buf);
     assert_eq!(hasher.finalize()[..], samedata_hash[..]);
+}
+
+
+pub fn open_file(fname: &'static str) -> File {
+    let path_base : PathBuf = PathBuf::from("test_resources/");
+    let fpath = path_base.join(fname);
+    assert!(fpath.exists());
+    File::open(fpath).unwrap()
 }
