--- tests/mod.rs
@@ -1,2 +1,0 @@
-#[allow(dead_code)]
-pub mod utils;
